Align with PR #3660
Align with Refactor remote services 27x #3657 / restored geonode.celery to geonode.celery_app #3531

 - Update dependencies and fix test cases

Fix tests

Fix test cases

+++ geonode/services/signals.py
@@ -36,10 +36,11 @@ def remove_harvest_job(sender, **kwargs):
     """Remove a Layer's harvest job so that it may be re-imported later."""
     layer = kwargs["instance"]
     if layer.remote_service is not None:
-        job = HarvestJob.objects.filter(resource_id=layer.name).get(
-            service=layer.remote_service)
-        logger.debug("job: {}".format(job.id))
-        job.delete()
+        if HarvestJob.objects.filter(resource_id=layer.alternate):
+            job = HarvestJob.objects.filter(resource_id=layer.name).get(
+                service=layer.remote_service)
+            logger.debug("job: {}".format(job.id))
+            job.delete()
     else:
         pass  # layer was not harvested from a service, we've nothing to do
